ECM-141: implemented flight connections service * created dedicated packages for airline, airport and connections services * modified name of user provided service * reformatted code * added embedded mongoDB for sake of integration test * removed manifest.yml from flightinformation folder
