--- services/flightinformation/manifest.yml
@@ -1,9 +1,0 @@
-# Configuration file for Cloud Foundry, see https://docs.cloudfoundry.org/devguide/deploy-apps/manifest.html
-applications:
-- name: flightservice
-  path: target/flightinformation-0.0.1-SNAPSHOT.jar
-  buildpack: https://github.com/cloudfoundry/java-buildpack.git#v3.10
-  memory: 512M
-  disk_quota: 256M
-  services:
-    - Educama_AtlasMongoDB
