+++ services/flightinformation/src/main/java/org/educama/App.java
@@ -1,0 +1,14 @@
+package org.educama;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.web.support.SpringBootServletInitializer;
+
+@SpringBootApplication
+public class App extends SpringBootServletInitializer {
+
+    public static void main(String[] args) {
+        SpringApplication.run(App.class, args);
+    }
+
+}
