+++ services/flightinformation/src/main/java/org/educama/airline/repository/AirlineRepository.java
@@ -1,0 +1,20 @@
+package org.educama.airline.repository;
+
+import org.educama.airline.model.Airline;
+import org.springframework.data.mongodb.repository.MongoRepository;
+
+import java.util.List;
+
+public interface AirlineRepository extends MongoRepository<Airline, String> {
+    public Airline findByName(String name);
+
+    public List<Airline> findByIataCode(String iataCode);
+
+    public List<Airline> findByIcaoCode(String icaoCode);
+
+    public List<Airline> findByNameLike(String name);
+
+    public List<Airline> findByIataCodeLike(String iataCode);
+
+    public List<Airline> findByCallSignLike(String callSign);
+}
