+++ services/flightinformation/src/main/java/org/educama/flightconnection/controller/ConnectionController.java
@@ -1,0 +1,48 @@
+package org.educama.flightconnection.controller;
+
+import org.educama.flightconnection.businessservice.ConnectionBusinessService;
+import org.educama.flightconnection.model.Connection;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
+
+import java.io.IOException;
+import java.util.List;
+
+@RestController
+public class ConnectionController {
+
+    ConnectionBusinessService connectionBusinessService;
+
+    @Autowired
+    public ConnectionController(ConnectionBusinessService connectionBusinessService) {
+        this.connectionBusinessService = connectionBusinessService;
+
+    }
+
+    /**
+     * Retrieves all existing connections  from a source to a destination airport.
+     * Retrieves all connection from a source (resp. destination) airport if only the source (resp. the destination) is specified.
+     *
+     * @param sourceAirportIata      the IATA code of the source airport
+     * @param destinationAirportIata the IATA code of the destination airport
+     * @return The available flight connections.
+     */
+    @RequestMapping(value = "/connections")
+    public List<Connection> getAllConnectionFromSourceToDestination(@RequestParam(value = "from") String sourceAirportIata, @RequestParam(value = "to") String destinationAirportIata) {
+        return connectionBusinessService.findAllConnectionsFromSourceToDestionation(sourceAirportIata, destinationAirportIata);
+    }
+
+
+    /**
+     * Replaces the content of the flight connection database with the content of the CSV file
+     * the data contained in the csv File.
+     *
+     * @param file the import file containing the flight connections data
+     */
+    @RequestMapping(value = "/connections/import/csv", method = RequestMethod.POST)
+    public @ResponseBody
+    void importConnections(@RequestParam("file") MultipartFile file) throws IOException {
+        connectionBusinessService.clearAndImportConnections(file);
+    }
+}
