--- services/flightinformation/src/main/java/org/educama/services/flightinformation/App.java
@@ -1,13 +1,0 @@
-package org.educama.services.flightinformation;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.web.support.SpringBootServletInitializer;
-
-@SpringBootApplication
-public class App extends SpringBootServletInitializer {
-
-	public static void main(String[] args) {
-		SpringApplication.run(App.class, args);
-	}
-}
