+++ services/flightinformation/src/main/resources/application.properties
@@ -1,9 +1,4 @@
 # File size limit
 spring.http.multipart.max-file-size=5MB
-
 # Total request size for a multipart/form-data
 multipart.maxRequestSize=10MB
-
-
-
-
