+++ services/flightinformation/src/main/resources/templates/csvUpload.html
@@ -16,38 +16,70 @@ float: left;
 margin: 10px 0 0 10px;
 width:100%;
 }
+
+
+
+
+
 </style>
 <body>
 <div class="main">
-	<div>
-		This is a little snippet for uploading csv file to the flightinfo service.
-	</div>
+    <div>
+        This is a little snippet for uploading csv file to the flightinfo service.
+    </div>
 
 
-	<div class="chart">
-		<h2>Upload Airports Data</h2>
-		The csv is posted as multipart data form to:<i> /airports/import/csv</i>
-		<form method="POST" enctype="multipart/form-data" action="airports/import/csv">
-			<table>
-				<tr>Please select the csv file</tr>
-				<tr><td><input type="file" name="file" /></td></tr>
-				<tr><td><input type="submit" value="Upload" /></td></tr>
-			</table>
-		</form>
-	</div>
+    <div class="chart">
+        <h2>Upload Airports Data</h2>
+        The csv is posted as multipart data form to:<i> /airports/import/csv</i>
+        <form method="POST" enctype="multipart/form-data" action="airports/import/csv">
+            <table>
+                <tr>Please select the csv file</tr>
+                <tr>
+                    <td><input type="file" name="file"/></td>
+                </tr>
+                <tr>
+                    <td><input type="submit" value="Upload"/></td>
+                </tr>
+            </table>
+        </form>
+    </div>
 
 
-	<div class="chart">
-		<h2>Upload Airlines Data</h2>
-		The csv is posted as multipart data form to:<i> /airlines/import/csv</i>
-		<form method="POST" enctype="multipart/form-data" action="airlines/import/csv">
-			<table>
-				<tr>Please select the csv file</tr>
-				<tr><td><input type="file" name="file" /></td></tr>
-				<tr><td><input type="submit" value="Upload" /></td></tr>
-			</table>
-		</form>
-	</div>
+    <div class="chart">
+        <h2>Upload Airlines Data</h2>
+
+        The csv is posted as multipart data form to:<i> /airlines/import/csv</i>
+        <form method="POST" enctype="multipart/form-data" action="airlines/import/csv">
+            <table>
+                <tr>Please select the csv file</tr>
+                <tr>
+                    <td><input type="file" name="file"/></td>
+                </tr>
+                <tr>
+                    <td><input type="submit" value="Upload"/></td>
+                </tr>
+            </table>
+        </form>
+    </div>
+
+
+    <div class="chart">
+        <h2>Upload flight connections Data</h2>
+        The csv is posted as multipart data form to:<i> /connections/import/csv</i>
+        <form method="POST" enctype="multipart/form-data" action="connections/import/csv">
+            <table>
+                <tr>Please select the csv file</tr>
+                <tr>
+                    <td><input type="file" name="file"/></td>
+                </tr>
+                <tr>
+                    <td><input type="submit" value="Upload"/></td>
+                </tr>
+            </table>
+        </form>
+    </div>
+
 </div>
 
 </body>
